fullanon/requirements.txt
# This file collects all required third-party applications that are needed
# to run this project. Later you can install all these apps in a row
# using pip. Example::
#
#     pip install -U -r requirements.txt

pydicom
pillow
numpy